requirements.txt
torch
torchvision
ultralytics
torchreid
scikit-learn
Pillow
numpy
opencv-python
